requirements.txt
fastapi
python-dotenv
pymysql
pillow
pytest
aiomysql
sqlalchemy
alembic
aiofiles
python-multipart
gunicorn
uvicorn
cryptography